--- .idea/shelf/Uncommitted_changes_before_Checkout_at_11_18_2020_3_37_PM_[Default_Changelist]/shelved.patch
@@ -1,56 +1,0 @@
-Index: build.gradle
-IDEA additional info:
-Subsystem: com.intellij.openapi.diff.impl.patch.BaseRevisionTextPatchEP
-<+>// Top-level build file where you can add configuration options common to all sub-projects/modules.\r\nbuildscript {\r\n    repositories {\r\n        google()\r\n        jcenter()\r\n    }\r\n    dependencies {\r\n        classpath 'com.android.tools.build:gradle:4.1.1'\r\n\r\n        // NOTE: Do not place your application dependencies here; they belong\r\n        // in the individual module build.gradle files\r\n    }\r\n}\r\n\r\nallprojects {\r\n    repositories {\r\n        google()\r\n        jcenter()\r\n    }\r\n}\r\n\r\ntask clean(type: Delete) {\r\n    delete rootProject.buildDir\r\n}
-Subsystem: com.intellij.openapi.diff.impl.patch.CharsetEP
-<+>UTF-8
-===================================================================
---- build.gradle	(revision 95a0c137789d6367aa7ec21b7d9894ec72fac17b)
-+++ build.gradle	(date 1605623169207)
-@@ -16,6 +16,7 @@
-     repositories {
-         google()
-         jcenter()
-+        mavenCentral()
-     }
- }
- 
-Index: .idea/vcs.xml
-IDEA additional info:
-Subsystem: com.intellij.openapi.diff.impl.patch.BaseRevisionTextPatchEP
-<+><?xml version=\"1.0\" encoding=\"UTF-8\"?>\r\n<project version=\"4\">\r\n  <component name=\"VcsDirectoryMappings\">\r\n  <mapping directory=\"$PROJECT_DIR$\" vcs=\"Git\"\r\n  </component>\r\n</project>
-Subsystem: com.intellij.openapi.diff.impl.patch.CharsetEP
-<+>UTF-8
-===================================================================
---- .idea/vcs.xml	(revision 95a0c137789d6367aa7ec21b7d9894ec72fac17b)
-+++ .idea/vcs.xml	(date 1605620135312)
-@@ -1,6 +1,6 @@
- <?xml version="1.0" encoding="UTF-8"?>
- <project version="4">
-   <component name="VcsDirectoryMappings">
--  <mapping directory="$PROJECT_DIR$" vcs="Git"
-+    <mapping directory="" vcs="Git" />
-   </component>
- </project>
-\ No newline at end of file
-Index: .idea/jarRepositories.xml
-IDEA additional info:
-Subsystem: com.intellij.openapi.diff.impl.patch.BaseRevisionTextPatchEP
-<+><?xml version=\"1.0\" encoding=\"UTF-8\"?>\r\n<project version=\"4\">\r\n  <component name=\"RemoteRepositoriesConfiguration\">\r\n    <remote-repository>\r\n      <option name=\"id\" value=\"central\" />\r\n      <option name=\"name\" value=\"Maven Central repository\" />\r\n      <option name=\"url\" value=\"https://repo1.maven.org/maven2\" />\r\n    </remote-repository>\r\n    <remote-repository>\r\n      <option name=\"id\" value=\"jboss.community\" />\r\n      <option name=\"name\" value=\"JBoss Community repository\" />\r\n      <option name=\"url\" value=\"https://repository.jboss.org/nexus/content/repositories/public/\" />\r\n    </remote-repository>\r\n    <remote-repository>\r\n      <option name=\"id\" value=\"BintrayJCenter\" />\r\n      <option name=\"name\" value=\"BintrayJCenter\" />\r\n      <option name=\"url\" value=\"https://jcenter.bintray.com/\" />\r\n    </remote-repository>\r\n    <remote-repository>\r\n      <option name=\"id\" value=\"Google\" />\r\n      <option name=\"name\" value=\"Google\" />\r\n      <option name=\"url\" value=\"https://dl.google.com/dl/android/maven2/\" />\r\n    </remote-repository>\r\n  </component>\r\n</project>
-Subsystem: com.intellij.openapi.diff.impl.patch.CharsetEP
-<+>UTF-8
-===================================================================
---- .idea/jarRepositories.xml	(revision 95a0c137789d6367aa7ec21b7d9894ec72fac17b)
-+++ .idea/jarRepositories.xml	(date 1605625558738)
-@@ -21,5 +21,10 @@
-       <option name="name" value="Google" />
-       <option name="url" value="https://dl.google.com/dl/android/maven2/" />
-     </remote-repository>
-+    <remote-repository>
-+      <option name="id" value="MavenRepo" />
-+      <option name="name" value="MavenRepo" />
-+      <option name="url" value="https://repo.maven.apache.org/maven2/" />
-+    </remote-repository>
-   </component>
- </project>
-\ No newline at end of file
