--- .idea/shelf/Uncommitted_changes_before_Checkout_at_11_18_2020_3_37_PM__Default_Changelist_.xml
@@ -1,4 +1,0 @@
-<changelist name="Uncommitted_changes_before_Checkout_at_11_18_2020_3_37_PM_[Default_Changelist]" date="1605688663454" recycled="true" deleted="true">
-  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_11_18_2020_3_37_PM_[Default_Changelist]/shelved.patch" />
-  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 11/18/2020 3:37 PM [Default Changelist]" />
-</changelist>
