--- .idea/shelf/Uncommitted_changes_before_Checkout_at_11_18_2020_8_04_PM__Default_Changelist_.xml
@@ -1,4 +1,0 @@
-<changelist name="Uncommitted_changes_before_Checkout_at_11_18_2020_8_04_PM_[Default_Changelist]" date="1605704708876" recycled="true" deleted="true">
-  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_11_18_2020_8_04_PM_[Default_Changelist]/shelved.patch" />
-  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 11/18/2020 8:04 PM [Default Changelist]" />
-</changelist>
