+++ app/src/main/java/com/example/cuciin_android/activity/modul/login/LoginPresenter.java
@@ -29,9 +29,9 @@ public class LoginPresenter implements LoginContract.Presenter{
     public void start(){
         UtilProvider.initUserSession(this.context);
 
-        if(UtilProvider.getUserSessionUtil().getSession() != null){
-            view.gotoDashboard();   //jika sudah login langsung masuk dashboard
-        }
+//        if(UtilProvider.getUserSessionUtil().getSession() != null){
+//            view.gotoDashboard();   //jika sudah login langsung masuk dashboard
+//        }
     }
 
     public int onLogin(User user){
@@ -39,7 +39,7 @@ public class LoginPresenter implements LoginContract.Presenter{
     }
 
     public void validateLogin(final User user, final Activity activity){
-        mApiService = UtilsApi.getAPIService();
+        mApiService = UtilsApi.getAPIServiceLocal();
         Call<LoginObj> call = mApiService.loginRequest(user.getUsername(), user.getPassword());
         call.enqueue(new Callback<LoginObj>() {
             @Override
