+++ app/src/main/java/com/example/cuciin_android/activity/modul/nearby/NearbyFragment.java
@@ -3,6 +3,7 @@ package com.example.cuciin_android.activity.modul.nearby;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -71,7 +72,6 @@ public class NearbyFragment extends BaseFragment<NearbyActivity, NearbyContract.
         ((RecycleViewAdapterNearby) mAdapter).setOnItemClickListener(new RecycleViewAdapterNearby.MyClickListener() {
             @Override
             public void onItemClick(int position, View view) {
-                //int id = listOutlet.get(position).getId();
                 Log.d("Dashboard", " >>>> " + position);
                 mPresenter.orderItem(activity, listOutlet.get(position));
             }
