+++ app/src/main/java/com/example/cuciin_android/activity/modul/order/OrderContract.java
@@ -5,6 +5,7 @@ import android.content.Intent;
 
 import com.example.cuciin_android.base.BasePresenter;
 import com.example.cuciin_android.base.BaseView;
+import com.example.cuciin_android.data.model.DataLaundryType;
 import com.example.cuciin_android.data.model.LaundryType;
 import com.example.cuciin_android.data.model.outlet.DataOutletObj;
 
@@ -19,7 +20,8 @@ public interface OrderContract {
 
     interface Presenter extends BasePresenter {
         int[] setData(int length);
-        void addTransaction(Activity activity, DataOutletObj outletObj, double[] amount, List<LaundryType> laundryType);
+        void addTransaction(Activity activity, DataOutletObj outletObj, int[] amount, LaundryType laundryType);
+        int hitungHarga(double[] amount, double[] price);
         //ArrayList<LaundryType> getDataLaundryType(JsonArray laundryType);
     }
 }
