+++ app/src/main/java/com/example/cuciin_android/activity/modul/order/OrderFragment.java
@@ -2,6 +2,7 @@
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -39,7 +40,7 @@ import java.util.List;
     private Button buttonOrder;
     int[] id;
 
-    public OrderFragment(DataOutletObj dataOutletObj, LaundryType laundryType, LoginObj loginObj) {
+    public OrderFragment(DataOutletObj dataOutletObj, LaundryType laundryType) {
         this.outletObj = dataOutletObj;
         this.laundryType = laundryType;
         this.loginObj = loginObj;
@@ -78,7 +79,8 @@ import java.util.List;
         buttonOrder.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                //mPresenter.addTransaction();
+                //Log.d("outlet Obj Name", outletObj.getName());
+                mPresenter.addTransaction(activity, outletObj, ((RecycleViewAdapterLaundryType) mAdapter).getAmount(), laundryType);
             }
         });
