+++ app/src/main/java/com/example/cuciin_android/data/model/login/DataLoginObj.java
@@ -10,13 +10,13 @@ public class DataLoginObj implements Serializable {
     @SerializedName("token")
     @Expose
     private String token;
-    @SerializedName("name")
+    @SerializedName("id")
     @Expose
-    private String name;
+    private int id;
 
-    public DataLoginObj(String token, String name){
+    public DataLoginObj(String token, int id){
         this.token = token;
-        this.name = name;
+        this.id = id;
     }
 
     public String getToken() {
@@ -27,11 +27,11 @@ public class DataLoginObj implements Serializable {
         this.token = token;
     }
 
-    public String getName() {
-        return name;
+    public int getId() {
+        return id;
     }
 
-    public void setName(String name) {
-        this.name = name;
+    public void setId(int id) {
+        this.id = id;
     }
 }
