+++ app/src/main/java/com/example/cuciin_android/helper/RetrofitClient.java
@@ -1,5 +1,7 @@
 package com.example.cuciin_android.helper;
 
+import android.util.Log;
+
 import okhttp3.OkHttpClient;
 import okhttp3.logging.HttpLoggingInterceptor;
 import retrofit2.Retrofit;
@@ -7,9 +9,12 @@ import retrofit2.converter.gson.GsonConverterFactory;
 
 public class RetrofitClient {
     private static Retrofit retrofit = null;
+    private static String baseUrlRetrofit = null;
 
-    public static Retrofit getClient(String baseUrl){
-        if(retrofit == null){
+    public static Retrofit getClient(String baseUrl) {
+        Log.d("Retrofit URL", baseUrl);
+        if(!baseUrl.equals(baseUrlRetrofit)) {
+            baseUrlRetrofit = baseUrl;
             retrofit = new Retrofit.Builder()
                     .baseUrl(baseUrl)
                     .addConverterFactory(GsonConverterFactory.create())
